Major corrections to document
- Do not restate the requirements in the design document.
- Please write complete sentences with a subject, verb and period.

Overall, my comment is that there is no "design" in your Design Document. The content so far is more of a Project Plan.
Software Design must identify the components to be created or modified. Each design element must trace back to requirements. Overarching design such as creation of an abstract bae class for "tools" belongs in a Design Document.