Support imageio-openjpeg for JPEG2000 reading
The imageio-openjpeg doesn't implement raster reading, but PDFBox no longer relies on raster support, so just removing that check makes imageio-openjpeg work with PDFBox when the required native library is installed.

However, when the native library is not installed, imageio-openjpeg indicates that it's not usable by returning null as the image reader. This needs to be checked to avoid NPEs merely from this library being installed without its native component.